backend/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
fastapi>=0.63.0
uvicorn>=0.12.0
python-multipart>=0.0.5
pandas>=1.0.0
numpy>=1.19.0
isort>=5.0.0
Pillow>=8.0.0
tqdm>=4.50.0
matplotlib>=3.3.0
seaborn>=0.11.0
scikit-learn>=0.24.0
